+++ packages/cmr-client/ValidationError.js
@@ -1,0 +1,9 @@
+'use strict';
+
+class ValidationError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = this.constructor.name;
+  }
+}
+module.exports = ValidationError;