+++ packages/cmrjs/utils.js
@@ -1,30 +1,41 @@
 const got = require('got');
 const _get = require('lodash.get');
 const publicIp = require('public-ip');
-const xml2js = require('xml2js');
-const { createErrorType } = require('@cumulus/common/errors');
+const { deprecate } = require('@cumulus/common/util');
+const { ValidationError } = require('@cumulus/cmr-client');
 
-const ValidationError = createErrorType('ValidationError');
+// getUrl is not part of the public cmr-client API
+const getUrl = require('@cumulus/cmr-client/getUrl');
+// validate is not part of the public cmr-client API
+const { validate } = require('@cumulus/cmr-client/ingestConcept');
+
+const xmlParseOptions = {
+  ignoreAttrs: true,
+  mergeAttrs: true,
+  explicitArray: false
+};
 
 /**
  * Returns the environment specific identifier for the input cmr environment.
+ * @deprecated
  * @param {string} env - cmr environment ['OPS', 'SIT', 'UAT']
  * @returns {string} - value to use to build correct cmr url for environment.
  */
 function hostId(env) {
-  const id = {
-    OPS: '',
-    SIT: 'sit',
-    UAT: 'uat'
-  };
-  return _get(id, env, 'uat');
+  deprecate('@cumulus/cmrjs/utils#hostId', '1.11.3');
+  return _get(
+    { OPS: '', SIT: 'sit', UAT: 'uat' },
+    env,
+    'uat'
+  );
 }
-
 
 /**
  * Determines the appropriate CMR host endpoint based on a given
  * value for CMR_ENVIRONMENT environment variable. Defaults
  * to the uat cmr
+ *
+ * @deprecated
  *
  * @param {Object} environment - process env like object
  * @param {string} environment.CMR_ENVIRONMENT - [optional] CMR environment to
@@ -35,6 +46,7 @@ function hostId(env) {
  * @returns {string} the cmr host address
  */
 function getHost(environment = process.env) {
+  deprecate('@cumulus/cmrjs/utils#getHost', '1.11.3');
   const env = environment.CMR_ENVIRONMENT;
   if (environment.CMR_HOST) return environment.CMR_HOST;
 
@@ -42,64 +54,18 @@ function getHost(environment = process.env) {
   return host;
 }
 
-const xmlParseOptions = {
-  ignoreAttrs: true,
-  mergeAttrs: true,
-  explicitArray: false
-};
-
-/**
- * returns the full url for various cmr services
- * based on the type passed, e.g. token, search, etc.
- *
- * @param {string} type - the type of the service, e.g. token, search
- * @param {string} cmrProvider - the CMR provider id
- * @returns {string} the cmr url
- */
-function getUrl(type, cmrProvider) {
-  let url;
-  const host = getHost();
-  const env = process.env.CMR_ENVIRONMENT;
-  const provider = cmrProvider;
-
-  switch (type) {
-  case 'token':
-    if (env === 'OPS') {
-      url = 'https://api.echo.nasa.gov/echo-rest/tokens/';
-    }
-    else if (env === 'SIT') {
-      url = 'https://testbed.echo.nasa.gov/echo-rest/tokens/';
-    }
-    else {
-      url = 'https://api-test.echo.nasa.gov/echo-rest/tokens/';
-    }
-    break;
-  case 'search':
-    url = `https://${host}/search/`;
-    break;
-  case 'validate':
-    url = `https://${host}/ingest/providers/${provider}/validate/`;
-    break;
-  case 'ingest':
-    url = `https://${host}/ingest/providers/${provider}/`;
-    break;
-  default:
-    url = null;
-  }
-
-  return url;
-}
-
-
 /**
  * Find the UMM version as a decimal string.
  * If a version cannot be found on the input object
  * version 1.4 is assumed and returned.
  *
+ * @deprecated
+ *
  * @param {Object} umm - UMM metadata object
  * @returns {string} UMM version for the given object
  */
 function ummVersion(umm) {
+  deprecate('@cumulus/cmrjs/utils#ummVersion', '1.11.3');
   return _get(umm, 'MetadataSpecification.Version', '1.4');
 }
 
@@ -107,6 +73,7 @@ function ummVersion(umm) {
  * Transform UMM version number to metadata format string.
  *
  * @param {string} versionNumber - UMM version string in decimal format (e.g. 1.4)
+ * @param {string} [ummFormat='json'] - [optional] UMM format, defaults to 'json'
  * @returns {string} UMM-G metadata format string (e.g. umm_json_v1_4)
  */
 function ummVersionToMetadataFormat(versionNumber, ummFormat = 'json') {
@@ -114,48 +81,10 @@ function ummVersionToMetadataFormat(versionNumber, ummFormat = 'json') {
 }
 
 /**
- * Posts a given xml string to the validate endpoint of the CMR
- * and returns the results
- *
- * @param {string} type - service type
- * @param {string} xml - the xml document
- * @param {string} identifier - the document identifier
- * @param {string} provider - the CMR provider
- * @returns {Promise.<boolean>} returns true if the document is valid
- */
-async function validate(type, xml, identifier, provider) {
-  let result;
-  try {
-    result = await got.post(`${getUrl('validate', provider)}${type}/${identifier}`, {
-      body: xml,
-      headers: {
-        'Content-type': 'application/echo10+xml'
-      }
-    });
-
-    if (result.statusCode === 200) {
-      return true;
-    }
-  }
-  catch (e) {
-    result = e.response;
-  }
-
-  const parsed = await new Promise((resolve, reject) => {
-    xml2js.parseString(result.body, xmlParseOptions, (err, res) => {
-      if (err) reject(err);
-      resolve(res);
-    });
-  });
-
-  throw new ValidationError(
-    `Validation was not successful, CMR error message: ${JSON.stringify(parsed.errors.error)}`
-  );
-}
-
-/**
  * Posts a given xml string to the validate endpoint of CMR
  * and promises true of valid.
+ *
+ * @deprecated
  *
  * @param {string} ummMetadata - the UMM object
  * @param {string} identifier - the document identifier
@@ -163,11 +92,12 @@ async function validate(type, xml, identifier, provider) {
  * @returns {Promise<boolean>} returns true if the document is valid
  */
 async function validateUMMG(ummMetadata, identifier, provider) {
+  deprecate('@cumulus/cmrjs/utils#validateUMMG', '1.11.3');
   const version = ummVersion(ummMetadata);
   let result;
 
   try {
-    result = await got.post(`${getUrl('validate', provider)}granule/${identifier}`, {
+    result = await got.post(`${getUrl('validate', provider, process.env.CMR_ENVIRONMENT)}granule/${identifier}`, {
       json: true,
       body: ummMetadata,
       headers: {
@@ -197,9 +127,12 @@ async function validateUMMG(ummMetadata, identifier, provider) {
  * static string if not defined. The value returned should be a valid IP address or
  * else the request for a CMR token will fail.
  *
+ * @deprecated
+ *
  * @returns {string} IP address
  */
 async function getIp() {
+  deprecate('@cumulus/cmrjs/utils#getIp', '1.11.3');
   return publicIp.v4()
     .catch((err) => {
       if (err.message === 'Query timed out') {
@@ -213,6 +146,8 @@ async function getIp() {
 /**
  * Returns a valid a CMR token
  *
+ * @deprecated
+ *
  * @param {string} cmrProvider - the CMR provider id
  * @param {string} clientId - the CMR clientId
  * @param {string} username - CMR username
@@ -220,12 +155,18 @@ async function getIp() {
  * @returns {Promise.<string>} the token
  */
 async function updateToken(cmrProvider, clientId, username, password) {
+  deprecate('@cumulus/cmrjs/utils#updateToken', '1.11.3');
+  if (!cmrProvider) throw new Error('cmrProvider is required.');
+  if (!clientId) throw new Error('clientId is required.');
+  if (!username) throw new Error('username is required.');
+  if (!password) throw new Error('password is required.');
+
   // Update the saved ECHO token
   // for info on how to add collections to CMR: https://cmr.earthdata.nasa.gov/ingest/site/ingest_api_docs.html#validate-collection
   let response;
 
   try {
-    response = await got.post(getUrl('token'), {
+    response = await got.post(getUrl('token', null, process.env.CMR_ENVIRONMENT), {
       json: true,
       body: {
         token: {